change requires to force current directory
In each of the files the require() and include() need to be updated due to changes in PHP and the DB directory in the the php default include folder. another option would be to change the include path but UserPie is not really set up to have a strict include path set.